Moving URL helpers into separate file.
