+++ resources/assets/components/App.js
@@ -6,7 +6,7 @@ import { ApolloProvider } from '@apollo/react-common';
 import { Provider as ReduxProvider } from 'react-redux';
 import { Router, Route, Switch } from 'react-router-dom';
 
-import { query } from '../helpers';
+import { query } from '../helpers/url';
 import graphqlClient from '../graphql';
 import ErrorPage from './pages/ErrorPage';
 import AuthGate from './utilities/AuthGate';
