+++ resources/assets/helpers/twitter.js
@@ -1,4 +1,5 @@
-import { makeUrl, openDialog } from '.';
+import { openDialog } from '.';
+import { makeUrl } from './url';
 import { EVENT_CATEGORIES, trackAnalyticsEvent } from './analytics';
 
 /**
